change batteryLevel type float with int

+++ app/src/main/java/com/rorono/maincomponentsofandroid/BroadcastReceiver.kt
@@ -12,18 +12,18 @@ import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import com.rorono.maincomponentsofandroid.screens.ContentProviderFragment
 
 class BroadcastReceiver : BroadcastReceiver() {
-    var batteryLevel: Float = 0.0f
+    var batteryLevel: Int = 0
 
     override fun onReceive(context: Context?, intent: Intent?) {
         val level = intent?.getIntExtra(BatteryManager.EXTRA_LEVEL, -1)
         val scale = intent?.getIntExtra(BatteryManager.EXTRA_SCALE, -1)
 
         if (level != null && scale != null) {
-            batteryLevel = level * 100 / scale.toFloat()
+            batteryLevel = level * 100 / scale
         }
     }
 
-    fun checkPhoneChargeLevel(): Float {
+    fun checkPhoneChargeLevel(): Int {
         return batteryLevel
     }
 }
